sort out app disconnection. move state appropriately
