+++ packages/react-app/src/views/MintingPage.jsx
@@ -1,9 +1,8 @@
-import React, { useReducer, useState, useContext, useEffect } from 'react'
+import React, { useReducer } from 'react'
 import Box from '@mui/material/Box'
 import Typography from '@mui/material/Typography'
 import MintingPageCard from '../components/MintingPageCard'
 import MintingActions from 'components/MintingActions'
-import { BadgeContext } from 'contexts/BadgeContext'
 import { useTheme } from '@mui/material/styles'
 import useMediaQuery from '@mui/material/useMediaQuery'
 import externalContracts from 'contracts/external_contracts'
